Add subject_id different from sample name validation (#3503) (patch)
### Added

- Validation ensuring that the sample name and subject id can not be the same.
