+++ cg/services/order_validation_service/models/errors.py
@@ -131,7 +131,7 @@ class InvalidGenePanelsError(CaseError):
 
 
 class InvalidBufferError(CaseSampleError):
-    field: str = "buffer"
+    field: str = "elution_buffer"
     message: str = "The chosen buffer is not allowed when skipping reception control"
 
 
@@ -143,3 +143,8 @@ class RepeatedGenePanelsError(CaseError):
 class SubjectIdSameAsCaseNameError(CaseSampleError):
     field: str = "subject_id"
     message: str = "Subject id must be different from the case name"
+
+
+class SubjectIdSameAsSampleNameError(CaseSampleError):
+    field: str = "subject_id"
+    message: str = "Subject id must be different from the sample name"
