+++ tests/services/order_validation_service/test_inter_field_validators.py
@@ -1,6 +1,5 @@
 from cg.services.order_validation_service.models.errors import (
     ApplicationNotCompatibleError,
-    InvalidBufferError,
     OrderNameRequiredError,
     SubjectIdSameAsCaseNameError,
     TicketNumberRequiredError,
@@ -8,7 +7,6 @@ from cg.services.order_validation_service.models.errors import (
 from cg.services.order_validation_service.models.order import Order
 from cg.services.order_validation_service.validators.inter_field.rules import (
     validate_application_compatibility,
-    validate_buffers_are_allowed,
     validate_name_required_for_new_order,
     validate_ticket_number_required_if_connected,
 )
@@ -65,21 +63,6 @@ def test_application_is_incompatible(
 
     # THEN the error should be about the application compatiblity
     assert isinstance(errors[0], ApplicationNotCompatibleError)
-
-
-def test_elution_buffer_is_not_allowed(valid_order: TomteOrder):
-
-    # GIVEN an order with 'skip reception control' toggled but no buffers specfied
-    valid_order.skip_reception_control = True
-
-    # WHEN validating that the buffers conform to the 'skip reception control' requirements
-    errors = validate_buffers_are_allowed(valid_order)
-
-    # THEN an error should be returned
-    assert errors
-
-    # THEN the error should be about the buffer compatability
-    assert isinstance(errors[0], InvalidBufferError)
 
 
 def test_subject_ids_same_as_case_names_not_allowed(valid_order: TomteOrder):
